--- src/main/java/com/example/estudiantes/dao/TurmaDAO.java
@@ -1,4 +1,0 @@
-package com.example.estudiantes.dao;
-
-public class TurmaDAO {
-}
