--- src/main/java/com/example/estudiantes/dao/UsuarioDAO.java
@@ -1,4 +1,0 @@
-package com.example.estudiantes.dao;
-
-public class UsuarioDAO {
-}
