+++ src/main/java/com/sistema/estudiantes/dao/AlunoDAO.java
@@ -1,11 +1,4 @@
-package com.example.estudiantes.dao;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Date;
+package com.sistema.estudiantes.dao;
 
 public class AlunoDAO {
