+++ src/main/java/com/sistema/estudiantes/dao/UsuarioDAO.java
@@ -1,0 +1,4 @@
+package com.sistema.estudiantes.dao;
+
+public class UsuarioDAO {
+}
